requirements.txt
Flask
Flask-Login
SQLAlchemy
Werkzeug
face_recognition
opencv-python
numpy
requests
opencv-python
opencv-contrib-python
face_recognition
dlib
numpy
pillow
ttkbootstrap
python-dotenv
mysql-connector-python
requests
scipy
cryptography
